fix : Refactoring Service and Dto
