+++ src/main/java/ongjong/namanmoo/controller/FamilyController.java
@@ -1,16 +1,10 @@
 package ongjong.namanmoo.controller;
 
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import ongjong.namanmoo.domain.Family;
-import ongjong.namanmoo.domain.MemberRole;
-import ongjong.namanmoo.dto.family.FamilyMemberDto;
+import ongjong.namanmoo.dto.family.*;
 import ongjong.namanmoo.response.*;
-import ongjong.namanmoo.response.family.CreateFamilyResponse;
-import ongjong.namanmoo.response.family.FamilyInfoResponse;
-import ongjong.namanmoo.response.family.FamilyInviteResponse;
-import ongjong.namanmoo.response.family.JoinFamilyResponse;
 import ongjong.namanmoo.service.FamilyService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -74,23 +68,5 @@ public class FamilyController {
                 familyInfoResponse
         );
         return ResponseEntity.ok(response);
-    }
-
-    @Getter
-    public static class CreateFamilyRequest {
-        private int familySize;
-        private String familyName;
-        private String ownerRole;
-    }
-
-    @Getter
-    public static class JoinFamilyRequest {
-        private Long familyId;
-        private String role;
-    }
-
-    @Getter
-    public static class FamilyIdRequest {
-        private String familyId;
     }
 }
