+++ src/main/java/ongjong/namanmoo/domain/VoiceMessage.java
@@ -14,7 +14,7 @@ public class VoiceMessage {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long voiceId;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
