+++ src/main/java/ongjong/namanmoo/dto/challenge/SaveAnswerRequest.java
@@ -1,0 +1,11 @@
+package ongjong.namanmoo.dto.challenge;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class SaveAnswerRequest {
+    private Long challengeId;
+    private String answerContent;
+}
