+++ src/main/java/ongjong/namanmoo/dto/challenge/SaveChallengeRequest.java
@@ -1,0 +1,10 @@
+package ongjong.namanmoo.dto.challenge;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class SaveChallengeRequest {
+    private Long challengeDate;
+}
