+++ src/main/java/ongjong/namanmoo/dto/family/FamilyInfoResponse.java
@@ -1,0 +1,13 @@
+package ongjong.namanmoo.dto.family;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+public class FamilyInfoResponse {
+    private List<FamilyMemberDto> members;
+
+}
