+++ src/main/java/ongjong/namanmoo/dto/family/FamilyInviteResponse.java
@@ -1,14 +1,14 @@
-package ongjong.namanmoo.response.family;
+package ongjong.namanmoo.dto.family;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import ongjong.namanmoo.dto.family.FamilyMemberDto;
 
 import java.util.List;
 
 @Data
 @AllArgsConstructor
-public class FamilyInfoResponse {
+public class FamilyInviteResponse {
+    private String familyName;
+    private String familyId;
     private List<FamilyMemberDto> members;
-
 }
