+++ src/main/java/ongjong/namanmoo/dto/family/JoinFamilyRequest.java
@@ -1,0 +1,12 @@
+package ongjong.namanmoo.dto.family;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class JoinFamilyRequest {
+    private Long familyId;
+    private String role;
+}
