+++ src/main/java/ongjong/namanmoo/dto/family/JoinFamilyResponse.java
@@ -1,0 +1,10 @@
+package ongjong.namanmoo.dto.family;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class JoinFamilyResponse {
+    private String familyId;
+}
