+++ src/main/java/ongjong/namanmoo/repository/FamilyRepository.java
@@ -13,7 +13,6 @@ import java.util.Optional;
 
 public interface FamilyRepository extends JpaRepository<Family, Long> {
     Optional<Family> findByInviteCode(String inviteCode);
-//    Optional<Family> findByFamilyId(long familyId);
 
     boolean existsByInviteCode(String inviteCode);
 }
