+++ src/main/java/ongjong/namanmoo/service/MemberService.java
@@ -5,7 +5,6 @@ import ongjong.namanmoo.dto.member.LoginRequestDto;
 import ongjong.namanmoo.dto.member.MemberInfoDto;
 import ongjong.namanmoo.dto.member.MemberSignUpDto;
 import ongjong.namanmoo.dto.member.MemberUpdateDto;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.Optional;
@@ -32,6 +31,5 @@ public interface MemberService {
 
     MemberInfoDto getMyInfo() throws Exception;
 
-    @Transactional(readOnly = true)
     Member findMemberByLoginId() throws Exception;
 }
