+++ src/main/java/ongjong/namanmoo/service/MemberServiceImpl.java
@@ -108,6 +108,7 @@ public class MemberServiceImpl implements MemberService {
         return new MemberInfoDto(findMember);
     }
 
+    @Override
     @Transactional(readOnly = true)
     public Member findMemberByLoginId() throws Exception{
         return memberRepository.findByLoginId(SecurityUtil.getLoginLoginId()).orElseThrow(() -> new Exception("회원이 없습니다"));
